Moved video section from document to picture/media
